requirements.txt
langchain==0.1.16
langchain_community==0.0.33
langchain_openai==0.0.8
openai==1.12.0
tiktoken==0.6.0
mlflow==2.12.1
faiss-cpu
starlette
fastapi
watchfiles 
slowapi
gunicorn
waitress
load_dotenv